Changes to the file saving to use temp files rather than memory since the quest is limited.  Few other things to benefit the quest.

+++ QuestomAssets/BeatSaber/DifficultyBeatmap.cs
@@ -20,7 +20,7 @@ namespace QuestomAssets.BeatSaber
         public Single NoteJumpMovementSpeed { get; set; }
 
         [JsonProperty("_noteJumpStartBeatOffset")]
-        public int NoteJumpStartBeatOffset { get; set; }
+        public Single NoteJumpStartBeatOffset { get; set; }
 
         [JsonProperty("_beatmapData")]
         public BeatmapDataObject BeatmapData { get; set; }
